btrfs-progs: help: print multiple syntax schemas on separate lines
The help string for some commands could be split to more lines for
clarity, eg. as is now in the receive command. The 'btrfs help' listing
should indent all the lines properly, similar the command specific
help with "usage:'.

The syntax of the first help string line is to separate all command
usage schemas by "\n".

Signed-off-by: David Sterba <dsterba@suse.com>