btrfs-progs: add new flag to read superblock with mismatched fsids
New flag that mimics OPEN_CTREE_IGNORE_FSID_MISMATCH but only for
reading the superblock. It should be passed around to various helpers
like scan or mount checks as they'd fail before we'd get to the final
caller that can do something useful with the filesystem.

This will be used for an interrupted 'btrfstune -u'.

Note to __open_ctree_fd: the RECOVERY mode is not compatible with that
flag

Signed-off-by: David Sterba <dsterba@suse.com>